% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{construct_table_from_formulas}
\alias{construct_table_from_formulas}
\title{Restructure formulas to fit within a model table}
\usage{
construct_table_from_formulas(x, ...)
}
\arguments{
\item{x}{Vector of \code{fmls} objects}
}
\description{
Restructure formulas to fit within a model table
}
\keyword{internal}
