% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{describe}
\alias{describe}
\title{Describe attributes of a \code{tm} vector}
\usage{
describe(x, property)
}
\arguments{
\item{x}{A vector \code{tm} objects}

\item{property}{A character vector of the following attributes of a \code{tm}
object: role, side, label, group, description, type, distribution}
}
\value{
A list of \code{term = property} pairs, where the term is the name of the
element (e.g. could be the `role' of the term).
}
\description{
Describe attributes of a \code{tm} vector
}
\examples{
f <- .o(output) ~ .x(input) + .m(mediator) + random
t <- tm(f)
describe(t, "role")

}
