% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{df_reconstruct}
\alias{df_reconstruct}
\title{If objects are model tables, attributes are carried over to subset object}
\usage{
df_reconstruct(x, to)
}
\description{
If objects are model tables, attributes are carried over to subset object
}
\keyword{internal}
