% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula_helpers}
\alias{formula_helpers}
\alias{lhs}
\alias{rhs}
\alias{rhs.formula}
\alias{lhs.formula}
\title{Tools for working with formula-like objects}
\usage{
lhs(x, ...)

rhs(x, ...)

\method{rhs}{formula}(x, ...)

\method{lhs}{formula}(x, ...)
}
\arguments{
\item{x}{A formula-like object}

\item{...}{Arguments to be passed to or from other methods}
}
\value{
A \code{character} describing part of a \code{formula} or \code{fmls} object
}
\description{
Tools for working with formula-like objects
}
