% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{model_table_reconstructable}
\alias{model_table_reconstructable}
\title{Can \code{mdl_tbl} be reconstructed based on invariants?}
\usage{
model_table_reconstructable(x, to)
}
\arguments{
\item{x}{data frame that will have invariants checked}

\item{to}{the tibble subclass of \code{mdl_tbl} that would be reconstructed}
}
\description{
Can \code{mdl_tbl} be reconstructed based on invariants?
}
\keyword{internal}
