% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{update.tm}
\alias{update.tm}
\title{Update \code{tm} objects}
\usage{
\method{update}{tm}(object, ...)
}
\arguments{
\item{object}{A \code{tm} object}

\item{...}{A series of \code{field = term ~ value} pairs that represent the
attribute to be updated. Can have a value of \code{NA} if the goal is to
remove an attribute or property.}
}
\value{
A \code{tm} object with updated attributes
}
\description{
This updates properties or attributes of a \code{tm} vector. This only updates
objects that already exist.
}
