% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-table.R
\name{validate_model_table}
\alias{validate_model_table}
\title{Model table object validation}
\usage{
validate_model_table(x)
}
\arguments{
\item{x}{data frame that will have invariants checked}
}
\description{
Model table object validation
}
\keyword{internal}
